Phase 1: move the artifact catalog below both of its callers
`local_ml` imported `Variant` from the prose rewriter while the rewriter's
catalog deferred-imported `MODELS` and `model_dir` back out of `local_ml`. The
deferred imports made the cycle loadable; they did not make the ownership
sound. The generic half now sits below both:

  inference/local_models/catalog.py       ModelSpec, ModelVariantSpec,
                                          RuntimeKind, the complete MODELS
  inference/local_models/assets.py        paths, presence, download, delete,
                                          prune (+ the prune self-check)
  inference/local_models/dependencies.py  per-runtime extras + install hint
  inference/local_models/__init__.py      facade, and available()

`Variant` is now `ModelVariantSpec` at every use, with no transitional alias.
`runtime` is a Literal, so a typo fails at the definition site. The manifest
stays complete at import: prune_stale deletes every unclaimed basename, so a
lazily-registered spec is a 4.7 GB checkpoint deleted by an unrelated button.

local_ml.py keeps the in-process llama-cpp calls and re-exports the moved
surface under __all__, because workflows/toolkit.py publishes this module by
name as the workflow author's API. Its first paragraph now says what it is
NOT, since local_ml and local_models are one letter apart.

Test patch targets follow the callee, not the facade (plan §8.2): a patch on
a re-export leaves production reading the real disk. `available()` reaches
`dependencies.deps_ok` / `assets.present` through their modules for the same
reason, and api/routes/local_ml.py calls the owning modules directly — before
this, the retargeted test patches let a route reach hf_hub_download for real.
