+++ backend/inference/local_ml.py
@@ -1,18 +1,25 @@
-"""Shared in-process CPU local-ML scaffold.
+"""In-process local-ML inference: the ``llama-cpp-python`` half.
+
+THE OTHER HALF IS :mod:`backend.inference.local_models`, one letter away and
+genuinely different: that package owns model *artifacts* (the manifest, the
+download, the prune) and the supervised ``llama-server`` child. This module
+owns the calls Orb makes into a GGUF loaded inside this process, and it
+re-exports the artifact surface because ``workflows/toolkit.py`` hands this
+module to workflow authors under this name.
 
 Opt-in: needs the ``requirements-ml.txt`` extras (``llama-cpp-python`` +
 ``huggingface_hub``) and a GGUF on disk. Base Orb never imports either — the
 imports live inside functions so a stock install runs fine and each local-ML
 route just 503s.
 
-One registry (``MODELS``), one download path, and per-feature cached ``Llama``
-handles. ``autocomplete`` is a text-generation model (``create_completion``);
-the rest are ModernBERT sequence classifiers read through ``_head_logits`` and
-differing only in how they interpret those logits — ``ascore`` softmaxes two,
-``aclassify`` argmaxes 28, ``aclassify_pov`` marginalizes a 4x3 grid. A new model
-reuses the download/toggle/path plumbing for free, but its *inference* path is its
-own — generation and classification don't share a call. The ``available`` /
-``complete`` / ``build_prompt`` / ``ascore`` names are the routes' stable surface.
+``autocomplete`` is a text-generation model (``create_completion``); the rest
+are ModernBERT sequence classifiers read through ``_head_logits`` and differing
+only in how they interpret those logits — ``ascore`` softmaxes two,
+``aclassify`` argmaxes 28, ``aclassify_pov`` marginalizes a 4x3 grid. A new
+model reuses the download/toggle/path plumbing for free, but its *inference*
+path is its own — generation and classification don't share a call. The
+``available`` / ``complete`` / ``build_prompt`` / ``ascore`` names are the
+routes' stable surface.
 """
 
 from __future__ import annotations
@@ -21,133 +28,63 @@ import asyncio
 import atexit
 import math
 import os
-import sys
 from collections.abc import Mapping, Sequence
-from dataclasses import dataclass
 from typing import Any
 
 from ..core.text_segmentation import remove_quoted_spans, split_sentences
-from .prose_rewriter.catalog import Variant
+from .local_models import (
+    MODELS,
+    ModelSpec,
+    ModelVariantSpec,
+    available,
+    delete_model,
+    deps_ok,
+    download,
+    import_llama,
+    install_cmd,
+    model_dir,
+    present,
+    prune_stale,
+    resolve_path,
+    variant_path,
+    variant_present,
+    variant_spec,
+)
 
-_ROOT = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+#: Re-exported from :mod:`local_models` for callers that address this module by
+#: name — ``workflows/toolkit.py`` publishes it as the workflow author's API,
+#: and the Local ML routes and tests import from here. NOTE FOR TESTS: these
+#: are second bindings. Production code calls ``local_models``' own copies, so
+#: a monkeypatch belongs on the module that OWNS the name (``assets.download``,
+#: ``dependencies.deps_ok``), not on the re-export.
+__all__ = [
+    "GO_EMOTIONS",
+    "MODELS",
+    "POV_ROWS",
+    "ModelSpec",
+    "ModelVariantSpec",
+    "acomplete",
+    "aclassify",
+    "aclassify_pov",
+    "ascore",
+    "available",
+    "build_prompt",
+    "complete",
+    "delete_model",
+    "deps_ok",
+    "download",
+    "install_cmd",
+    "model_dir",
+    "pov_from_logits",
+    "pov_input",
+    "present",
+    "prune_stale",
+    "resolve_path",
+    "variant_path",
+    "variant_present",
+    "variant_spec",
+]
 
-
-@dataclass(frozen=True)
-class ModelSpec:
-    """One local-ML feature's weights, and how they are run.
-
-    ``runtime`` names the engine, because the two differ in what they need
-    installed: ``llama_cpp`` features are in-process and want the full
-    ``requirements-ml.txt``; ``llama_server`` features (the prose rewriter)
-    drive a child ``llama-server`` over HTTP and need only ``huggingface_hub``,
-    for the download. ``deps_ok(feature)`` reads this — a stock install with
-    just ``huggingface_hub`` can run the rewriter and nothing else.
-
-    ``variants`` is non-empty when the feature ships several interchangeable
-    checkpoints the user picks between. The spec's own ``filename`` then names
-    the default one, so every path that predates variants (``resolve_path``,
-    ``download``) keeps working unchanged.
-    """
-
-    repo_id: str
-    filename: str  # path *inside the HF repo* — upstream's layout, not ours
-    size_mb: int
-    revision: str  # pinned commit sha — a repo re-point can't swap the weights under us
-    runtime: str = "llama_cpp"  # llama_cpp (in-process) | llama_server (child process)
-    variants: tuple[Variant, ...] = ()
-
-    @property
-    def local_name(self) -> str:
-        """On-disk name under data/models/, always flat.
-
-        Upstream repos disagree about where a GGUF lives — root, ``gguf/``,
-        ``GGUF/`` — and mirroring that gave us a tree whose two case-variant
-        directories are ONE directory on macOS/Windows. Basenames must stay
-        unique across MODELS *and* across every spec's variants — a name two
-        specs both claim is one file two features would fight over, and a
-        variant name no spec claims is a file ``prune_stale`` deletes the next
-        time anything downloads. ``test_prose_rewriter_catalog`` asserts both.
-        """
-        return os.path.basename(self.filename)
-
-    def all_names(self) -> set[str]:
-        """Every basename this spec puts under data/models/ — the prune claim."""
-        return {self.local_name, *(v.local_name for v in self.variants)}
-
-
-# The two prose-rewriter repos, pinned. Named once because three variants
-# share them and a half-updated pin is a silently different model.
-_PROSE_1_7B_REPO = "chartreuse-verte/prose-rewriter-1.7b-v1.2"
-_PROSE_1_7B_REV = "9376423c82bf236c9cfc8d79fa1675fa104d4979"
-_PROSE_4B_REPO = "chartreuse-verte/prose-rewriter-4b-v1.2"
-_PROSE_4B_REV = "b73aa3290a81024389e06c9d525238071ccbe480"
-
-MODELS: dict[str, ModelSpec] = {
-    "autocomplete": ModelSpec(
-        repo_id="chartreuse-verte/orb-human-typeahead-1b-v2.2",
-        filename="GGUF/orb-human-typeahead-1b-v2.2-Q4_0.gguf",
-        size_mb=930,
-        revision="2e340db799eca2ef36ef80fc6938e40ab1ece111",
-    ),
-    "slop_classifier": ModelSpec(
-        repo_id="chartreuse-verte/ettin150m-purple-GGUF",
-        filename="ettin150m-purple-q8_0.gguf",
-        size_mb=161,
-        revision="125cf38d62e78b7091c23e6d523d805c7ec2f47e",
-    ),
-    "emotion_classifier": ModelSpec(
-        repo_id="chartreuse-verte/ettin-emotion-28-multilabel-68m",
-        filename="gguf/ettin-emotion-28ml-68m-q8_0.gguf",
-        size_mb=71,
-        revision="9f8d0100e45c133e713283499e55105f61d29118",
-    ),
-    "pov_classifier": ModelSpec(
-        repo_id="chartreuse-verte/ettin-povtense-17m",
-        filename="gguf/povtense-17m-q8_0.gguf",
-        size_mb=20,
-        revision="1245e55c47f9afc3d4938ef70f5228580228d899",
-    ),
-    # Not an in-process model: served by a child llama-server (see
-    # inference/prose_rewriter/). `filename`/`size_mb` name the default variant
-    # so the legacy single-file paths below keep working; the selector reads
-    # `variants`, and every basename here must also be claimed by prune_stale.
-    "prose_rewriter": ModelSpec(
-        repo_id=_PROSE_4B_REPO,
-        filename="GGUF/prose-rewriter-4b-v1.2-Q8_0.gguf",
-        size_mb=4694,
-        revision=_PROSE_4B_REV,
-        runtime="llama_server",
-        variants=(
-            Variant(
-                id="1.7b-q8",
-                label="1.7B · Q8_0",
-                detail="Fastest, good enough.",
-                repo_id=_PROSE_1_7B_REPO,
-                path="GGUF/prose-rewriter-1.7b-v1.2-Q8_0.gguf",
-                revision=_PROSE_1_7B_REV,
-                size_mb=2165,
-            ),
-            Variant(
-                id="4b-q4km",
-                label="4B · Q4_K_M",
-                detail="Medium quality.",
-                repo_id=_PROSE_4B_REPO,
-                path="GGUF/prose-rewriter-4b-v1.2-Q4_K_M.gguf",
-                revision=_PROSE_4B_REV,
-                size_mb=2716,
-            ),
-            Variant(
-                id="4b-q8",
-                label="4B · Q8_0",
-                detail="Best quality, invents the least.",
-                repo_id=_PROSE_4B_REPO,
-                path="GGUF/prose-rewriter-4b-v1.2-Q8_0.gguf",
-                revision=_PROSE_4B_REV,
-                size_mb=4694,
-            ),
-        ),
-    ),
-}
 
 # The 28 go-emotions labels in standard id2label order (neutral last, index 27).
 # Order MUST match the GGUF head's logit order — the classifier reads argmax(v[0:28])
@@ -223,171 +160,11 @@ def _lock(feature: str) -> asyncio.Lock:
     return _locks.setdefault(feature, asyncio.Lock())
 
 
-def model_dir() -> str:
-    d = os.path.join(_ROOT, "backend", "data", "models")
-    os.makedirs(d, exist_ok=True)
-    return d
-
-
-def resolve_path(feature: str) -> str:
-    """Where feature's GGUF lives: env override → data/models → repo root (back-compat)."""
-    if feature == "autocomplete":
-        env = os.environ.get("ORB_AUTOCOMPLETE_MODEL")
-        if env and os.path.exists(env):  # stale override must not hide a downloaded model
-            return env
-    spec = MODELS[feature]
-    for candidate in (
-        os.path.join(model_dir(), spec.local_name),  # flat — what download() writes
-        os.path.join(model_dir(), spec.filename),  # legacy: hf's mirror of the repo layout
-        os.path.join(_ROOT, spec.local_name),  # legacy: manual drop at repo root
-        os.path.join(_ROOT, spec.filename),  # legacy: mirrored drop at repo root
-    ):
-        if os.path.exists(candidate):
-            return candidate
-    return os.path.join(model_dir(), spec.local_name)  # absent: name the flat path in errors
-
-
-def _import_llama():
-    from llama_cpp import Llama  # noqa: PLC0415 — deferred so base Orb needs no ML deps
-
-    return Llama
-
-
-def _shell_quote(path: str) -> str:
-    """Quote only when needed — `C:\\Program Files\\...` breaks an unquoted paste."""
-    return f'"{path}"' if " " in path else path
-
-
-def install_cmd() -> str:
-    """Install command for THIS interpreter, fully qualified — a bare `pip` targets
-    whatever's on PATH, not the venv/uv env the server actually runs under, so the
-    extras land in the wrong Python and the button stays gray; and a bare
-    requirements filename only resolves if the shell happens to be cwd'd into the
-    repo, which a fresh cmd prompt is not."""
-    req = os.path.join(_ROOT, "requirements-ml.txt")
-    return f"{_shell_quote(sys.executable)} -m pip install -r {_shell_quote(req)}"
-
-
-def deps_ok(feature: str | None = None) -> tuple[bool, str]:
-    """Cheap check (no model load): are *feature*'s extras importable?
-
-    Per-runtime, because the features no longer share one answer. A
-    ``llama_server`` feature drives a child process over HTTP and needs only
-    ``huggingface_hub``, to fetch the weights; ``llama_cpp`` features run the
-    model in-process and need the binding too. ``feature=None`` keeps the
-    original whole-extras meaning, which is what the Local ML card's top-level
-    ``deps_ok`` (the grouped opt-in) is keyed on; every per-feature caller
-    passes a name.
-    """
-    runtime = MODELS[feature].runtime if feature in MODELS else "llama_cpp"
-    try:
-        if runtime == "llama_cpp":
-            _import_llama()
-        import huggingface_hub  # noqa: F401, PLC0415 — deferred; only needed for downloads
-    except Exception as e:  # ModuleNotFoundError or a broken build
-        return False, f"ML extras not installed ({e}); {install_cmd()}"
-    return True, ""
-
-
-def present(feature: str) -> bool:
-    """Is *feature* usable from disk?
-
-    For a variant-bearing spec that means ANY variant is downloaded — the
-    Settings card flips from "download something" to "pick one and enable" on
-    the first file, not on the default one.
-    """
-    spec = MODELS.get(feature)
-    if spec is not None and spec.variants:
-        return any(os.path.exists(os.path.join(model_dir(), v.local_name)) for v in spec.variants)
-    return os.path.exists(resolve_path(feature))
-
-
-def prune_stale(root: str | None = None) -> None:
-    """Delete any .gguf under data/models/ that no current MODELS spec claims.
-
-    Runs after every download so bumping a model (e.g. v2 typeahead) doesn't leave
-    the old weights eating disk. Only touches .gguf files — hf's .cache bookkeeping
-    and manual drops of other extensions are left alone.
-
-    Claim is by *basename*, not full path: comparing paths meant a model sitting in
-    a legacy mirrored subdir read as unclaimed and got deleted the moment any other
-    feature downloaded — and on a case-insensitive filesystem, where ``GGUF/`` and
-    ``gguf/`` are one directory, that fired on a model we had just fetched.
-    """
-    root = root or model_dir()
-    # Every basename a spec puts on disk, VARIANTS INCLUDED. A variant the claim
-    # set forgets is wiped the next time any feature downloads — 4.7 GB gone
-    # because an unrelated button was pressed.
-    keep = {name for s in MODELS.values() for name in s.all_names()}
-    walked: list[str] = []
-    for dirpath, dirs, files in os.walk(root):
-        dirs[:] = [d for d in dirs if d != ".cache"]  # hf's bookkeeping is its own business
-        for name in files:
-            if name.endswith(".gguf") and name not in keep:
-                os.remove(os.path.join(dirpath, name))
-        if dirpath != root:
-            walked.append(dirpath)
-    for d in reversed(walked):  # deepest first, so the emptied-out gguf//GGUF/ mirrors collapse
-        if not os.listdir(d):
-            os.rmdir(d)
-
-
-def variant_spec(feature: str, variant_id: str | None) -> tuple[str, str, str, str]:
-    """``(repo_id, path_in_repo, revision, local_name)`` for one download.
-
-    A ``variant_id`` names one of the spec's variants; ``None`` falls back to
-    the spec's own file, which is what every single-file feature uses.
-    """
-    spec = MODELS[feature]
-    if variant_id:
-        for v in spec.variants:
-            if v.id == variant_id:
-                return v.repo_id, v.path, v.revision, v.local_name
-        raise ValueError(f"Unknown variant {variant_id!r} for {feature!r}")
-    return spec.repo_id, spec.filename, spec.revision, spec.local_name
-
-
-def download(feature: str, variant: str | None = None) -> None:
-    """Fetch a GGUF into data/models/, then prune stale weights. Blocking; run in a thread."""
-    from huggingface_hub import hf_hub_download  # noqa: PLC0415 — deferred
-
-    repo_id, path, revision, local_name = variant_spec(feature, variant)
-    got = hf_hub_download(repo_id=repo_id, filename=path, revision=revision, local_dir=model_dir())
-    flat = os.path.join(model_dir(), local_name)
-    if os.path.normpath(got) != os.path.normpath(flat):
-        os.replace(got, flat)  # hf mirrors the repo's own gguf//GGUF/ nesting; we don't keep it
-    prune_stale()  # after fetch: new file lands before old ones go, so a failed download keeps the old model
-
-
-def delete_model(feature: str, variant: str | None = None) -> bool:
-    """Remove one downloaded GGUF. Returns whether a file was actually deleted.
-
-    Exists because the three rewriter variants are 9.6 GB combined and "find
-    the folder yourself" is not an acceptable only exit at that size.
-    """
-    _repo, _path, _rev, local_name = variant_spec(feature, variant)
-    target = os.path.join(model_dir(), local_name)
-    if not os.path.exists(target):
-        return False
-    os.remove(target)
-    return True
-
-
-def available(feature: str = "autocomplete") -> tuple[bool, str]:
-    """Feature readiness: extras installed AND this feature's model present."""
-    ok, reason = deps_ok(feature)
-    if not ok:
-        return False, reason
-    if not present(feature):
-        return False, f"model file not found: {resolve_path(feature)}"
-    return True, ""
-
-
 def _load_blocking(feature: str) -> None:
     if feature in _llamas or feature in _load_errors:
         return
     try:
-        Llama = _import_llama()
+        Llama = import_llama()
         _llamas[feature] = Llama(
             model_path=resolve_path(feature),
             n_ctx=1024,
@@ -688,26 +465,3 @@ if __name__ == "__main__":
     assert pov_input("") == "" and pov_input("   ") == ""
     assert pov_input("no terminal punctuation here") == "no terminal punctuation here"
     print("pov_input OK")
-
-    # Self-check for the destructive prune (temp dir; never touches real models).
-    import tempfile
-
-    with tempfile.TemporaryDirectory() as d:
-        keep = os.path.join(d, MODELS["autocomplete"].local_name)
-        open(keep, "w").close()
-        mirrored = os.path.join(d, MODELS["pov_classifier"].filename)  # legacy gguf/ nesting
-        os.makedirs(os.path.dirname(mirrored), exist_ok=True)
-        open(mirrored, "w").close()
-        stale = os.path.join(d, "old-granite-Q8_0.gguf")
-        open(stale, "w").close()
-        notes = os.path.join(d, "readme.txt")  # non-gguf must survive
-        open(notes, "w").close()
-        cached = os.path.join(d, ".cache", "huggingface", "download")
-        os.makedirs(cached)
-        prune_stale(d)
-        assert os.path.exists(keep), "current spec's gguf must be kept"
-        assert os.path.exists(mirrored), "a claimed gguf in a legacy subdir must survive"
-        assert not os.path.exists(stale), "unclaimed gguf must be removed"
-        assert os.path.exists(notes), "non-gguf must be left alone"
-        assert os.path.isdir(cached), "hf's .cache must be left alone, empty or not"
-    print("prune_stale OK")
