+++ tests/integration/test_character_expressions.py
@@ -80,8 +80,8 @@ async def test_upload_missing_card_404(client):
 
 
 async def test_classify_emotion_503_when_deps_absent(client, monkeypatch):
-    from backend.inference import local_ml
+    from backend.inference.local_models import dependencies
 
-    monkeypatch.setattr(local_ml, "deps_ok", lambda feature=None: (False, "extras not installed"))
+    monkeypatch.setattr(dependencies, "deps_ok", lambda feature=None: (False, "extras not installed"))
     resp = await client.post("/api/local-ml/classify-emotion", json={"text": "I am so happy!"})
     assert resp.status_code == 503
