+++ tests/unit/test_local_models_paths.py
@@ -16,20 +16,20 @@ from __future__ import annotations
 import os
 from pathlib import Path
 
-from backend.inference import local_ml
+from backend.inference.local_models import assets, dependencies
 from backend.inference.prose_rewriter import runtime
 
 ROOT = Path(__file__).resolve().parents[2]
 
 
 def test_model_and_binary_dirs_resolve_under_backend_data():
-    assert Path(local_ml.model_dir()) == ROOT / "backend" / "data" / "models"
+    assert Path(assets.model_dir()) == ROOT / "backend" / "data" / "models"
     assert Path(runtime.bin_dir()) == ROOT / "backend" / "data" / "llama-bin"
 
 
 def test_the_install_command_names_this_repos_requirements_file():
     """Same ``_ROOT``, different consumer: the pip line the Settings panel
     shows is pasted into a shell with no cwd in the repo."""
-    req = local_ml.install_cmd().split("-r ", 1)[1].strip('"')
+    req = dependencies.install_cmd().split("-r ", 1)[1].strip('"')
     assert Path(req) == ROOT / "requirements-ml.txt"
     assert os.path.isabs(req)
